sidebar: support indenting and shortening of names
Sidebar folder indenting and shortening only work on a mailbox's path.
However, the sidebar will always use a mailbox's name so indenting and
shortening does not work for 'virtual-mailboxes' or 'named-mailboxes'.

To support these operations, extend the indentation and shorten code to
use the name if it exists. This behavior is an extension of the current
behavior of names taking presidence.

Now the sidebar behaves more consistently across all mailbox types.

Closes #1594